AI_Agent/requirements.txt
PyPDF2
yfinance
transformers
torch
sentence-transformers
beautifulsoup4
requests
google-api-python-client
google-generativeai
